Move election models into own app.
